requirements.txt
tabulate>=0.9.0
python-dateutil>=2.8.2 